+++ datakit_gitlab/__init__.py
@@ -2,7 +2,7 @@
 
 __author__ = """Serdar Tumgoren"""
 __email__ = 'stumgoren@ap.org'
-__version__ = '0.1.0'
+__version__ = '0.2.0'
 
 from .commands.integrate import Integrate
 from .commands import issues