+++ HISTORY.rst
@@ -2,6 +2,13 @@
 History
 =======
 
-0.1.0 (TBD)
-------------------
+0.2.0
+-----
 
+* Port commands to use spaces rather than colons
+
+0.1.0
+------
+
+* Gitlab integrate
+* Gitlab issue add